[#11] Fix return value for apt_update
On the agent side apt_update, update the packages information and
returns the status of this operation (success and failure), while the
application side expect a list of outdated packages to be provided.

This list of outdated packages is available using the apt_checkupdates
action of the agent.

Fix this issue by calling the apt_checkupdates method and return it's
value if the update succeeds.